Fix modifier deactivation
Invalid `true` and `false` constants were use. In a sane language
this would have been a compile-time error. Additionally, the
deactivated modifier was being re-appended to the list of deactivated
modifers, causing an infinite loop. In a sane language, this would
have been a run-time error (concurrent modification of an interated
collection).