Start redesigning formats for better versioning
This redesigns the text format such that:

Assuming that a parser that supports major version m is receiving a
file of version m.n for any n:

  1. For header parsing, ignore and log any unknown header commands
     as warnings.
  2. For body parsing, ignore and log any unknown body section commands
     as warnings, and suppress the printing of any warnings until the
     next recognized body section command.

Sections must be ended with an explicit end. This drastically
simplifies parsing as a whole, and allows skipping entire sections
when those sections are unrecognized.

Affects #41