Continue redesigning formats for better versioning
This redesigns the binary format to use a sectioned or chunked format
similar to PNG. Sections begin with a size and a magic number. This
ensures that unrecognized data can be skipped for forwards and
backwards compatibility. Sections are 16-byte aligned: 32 or 64 byte
alignment is too wasteful.

This also changes schema and metadata identifiers to use reverse-DNS
64-character strings.

Affects #41
Fix #46
Fix #45