Use an explicit header size rather than an attribute offset
The total size of the header fields (minus the attributes) are declared
up front at the start of the header. This will prevent future additions
to the header from having to introduce their own "attribute_offset"
field variations.

This also rearranges the padding so that all header fields are
strictly aligned according to their type, and the header field total
is a multiple of 16.

Fix #49